Fix inaccurate mean speed if Heatmap was used for existing saves
If Heatmap was installed and used for existing saves instead of starting a new save, the average speed could be too high for nodes that were already occupied. Now these are ignored, which also means the visit counter and time spent modes are not updated for that node until a new train enters it.